requirements.txt
# Core Dependencies
click>=8.0.0
pyyaml>=6.0
requests>=2.28.0
pydantic
email-validator
jwt
psutil
python-dotenv>=0.20.0

# Web Interface
flask>=2.2.0
flask-restx>=1.0.0
flask-login>=0.6.0
flask-wtf>=1.0.0
jinja2>=3.0.0
werkzeug>=2.2.0

# Database Connectors
sqlalchemy>=1.4.0
pymongo>=4.2.0
elasticsearch-dsl>=7.4.0
neo4j>=5.0.0

# Task Scheduling
celery>=5.2.0
redis>=4.3.0
apscheduler>=3.9.0

# Security Modules
cryptography>=38.0.0
paramiko>=2.11.0
scapy>=2.4.5
pyopenssl>=22.0.0

# OSINT Modules
beautifulsoup4>=4.11.0
dnspython>=2.2.0
python-whois>=0.7.0
shodan>=1.28.0

# Visualization
matplotlib>=3.5.0
networkx>=2.8.0
plotly>=5.10.0

# Testing
pytest>=7.0.0
coverage>=6.4.0
pytest-cov>=3.0.0

# Development
black>=22.6.0
flake8>=5.0.0
mypy>=0.971
isort>=5.10.0
pre-commit>=2.20.0

# Documentation
sphinx>=5.1.0
sphinx-rtd-theme>=1.0.0
